requirements.txt
# Google Workspace API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# AI Model SDKs
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2

# Async support
aiohttp==3.9.3
asyncio==3.4.3

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3

# Logging and monitoring
python-json-logger==2.0.7

# Development tools (commented out for now - not needed for basic test)
# pytest==8.0.0
# pytest-asyncio==0.23.4
# black==24.1.1
# ruff==0.2.1
